rename deployment scripts (& update readme, package.json)
