+++ tasks/deployment/deploy-incentives-impl.ts
@@ -1,14 +1,12 @@
 import { task } from 'hardhat/config';
 import { DefenderRelaySigner, DefenderRelayProvider } from 'defender-relay-client/lib/ethers';
-import { deployStakedTokenIncentivesController } from '../../helpers/contracts-accessors';
+import { deployInitializableAdminUpgradeabilityProxy, deployStakedTokenIncentivesController } from '../../helpers/contracts-accessors';
 import { getDefenderRelaySigner } from '../../helpers/defender-utils';
-
-// Mainnet addresses
-// const STAKED_STARLAY = '0x4da27a545c0c5B758a6BA100e3a049001de870f5';
-// const STARLAY_SHORT_EXECUTOR = '0xee56e2b3d491590b5b31738cc34d5232f378a8d5';
+import { waitForTx } from '../../helpers/misc-utils';
+import { eContractid } from '../../helpers/types';
 
 // Shibuya (astar network) addresses
-const STAKED_STARLAY = '0x5d666338118763ca0cF6719F479491B76bc88131'; // StakedLay's (StakedTokenV2Rev3) proxy
+const STAKED_STARLAY = '0x3bA5A4b88331627236378000B50eAa186375c3FF'; // StakedLay's (StakedTokenV2Rev3) proxy
 const STARLAY_SHORT_EXECUTOR = '0x6543076E4315bd82129105890Bc49c18f496a528'; // PoolAdmin
 
 task('deploy-incentives-impl', 'Incentives controller implementation deployment').setAction(
@@ -16,6 +14,7 @@ task('deploy-incentives-impl', 'Incentives controller implementation deployment'
     _;
     await localBRE.run('set-DRE');
 
+    // setup deployer
     let deployer;
     if (process.env.DEFENDER_API_KEY && process.env.DEFENDER_SECRET_KEY) {
       const { signer } = await getDefenderRelaySigner();
@@ -25,13 +24,34 @@ task('deploy-incentives-impl', 'Incentives controller implementation deployment'
       deployer = signer;
     }
 
-    const incentives = await deployStakedTokenIncentivesController(
-      [STAKED_STARLAY], // TODO: reflect STARLAY_SHORT_EXECUTOR
-      false, // TODO: revert
-      deployer
+    console.log(`[StakedTokenIncentivesController] Starting deployment:`);
+    const impl = await deployStakedTokenIncentivesController(
+      [STAKED_STARLAY],
+      false // TODO: select necessity to verify
     );
-    console.log(`- Incentives implementation address ${incentives.address}`);
+    console.log(`  - Deployed implementation of ${eContractid.StakedTokenIncentivesController}`);
+    const proxy = await deployInitializableAdminUpgradeabilityProxy(false);  // TODO: select necessity to verify
+    console.log(`  - Deployed proxy of ${eContractid.StakedTokenIncentivesController}`);
+    const encodedParams = impl.interface.encodeFunctionData('initialize', [
+      STARLAY_SHORT_EXECUTOR
+    ]);
 
-    return incentives.address;
+    await waitForTx(
+      await proxy.functions['initialize(address,address,bytes)'](
+        impl.address,
+        await deployer.getAddress(),
+        encodedParams
+      )
+    );
+    console.log(`  - Initialized ${eContractid.StakedTokenIncentivesController} Proxy`);
+
+    console.log(`  - Finished PullRewardsIncentivesController deployment and initialization`);
+    console.log(`    - Proxy: ${proxy.address}`);
+    console.log(`    - Impl: ${impl.address}`);
+
+    return {
+      proxy: proxy.address,
+      implementation: impl.address
+    };
   }
 );
